+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - moved Adguard API to separate package — [@adguard/api](https://www.npmjs.com/package/@adguard/api)
 - updated `@adguard/extended-css` to [v2.0.51](https://github.com/AdguardTeam/ExtendedCss/blob/master/CHANGELOG.md#v2051)
+- updated `@adguard/tswebextension` to [v0.2.0](https://github.com/AdguardTeam/tsurlfilter/blob/master/packages/tswebextension/CHANGELOG.md#020---2023-05-23)